+++ lib/service/roco/z21_service.dart
@@ -40,13 +40,33 @@ int littleEndianData2uint32(Uint8List data) {
 }
 
 /// \todo document
-int data2locoAddress(Uint8List data) {
+int bigEndianData2locoAddress(Uint8List data) {
   return bigEndianData2uint16(data) & 0x3FFF;
 }
 
 /// \todo document
-int data2cvAddress(Uint8List data) {
+int bigEndianData2cvAddress(Uint8List data) {
   return bigEndianData2uint16(data) & 0x03FF;
+}
+
+/// \todo document
+int bigEndianLocoAddressMsb(int locoAddress) {
+  return (locoAddress >= 128 ? 0xC0 : 0x00) | ((locoAddress >> 8) & 0xFF);
+}
+
+/// \todo document
+int bigEndianLocoAddressLsb(int locoAddress) {
+  return (locoAddress >> 0) & 0xFF;
+}
+
+/// \todo document
+int littleEndianLocoAddressMsb(int locoAddress) {
+  return (locoAddress >> 0) & 0xFF;
+}
+
+/// \todo document
+int littleEndianLocoAddressLsb(int locoAddress) {
+  return (locoAddress >> 8) & 0xFF;
 }
 
 /// \todo document
@@ -457,7 +477,7 @@ class LanXCvResult implements Command {
   LanXCvResult({required this.cvAddress, required this.value});
 
   LanXCvResult.fromDataset(Uint8List dataset)
-      : cvAddress = data2cvAddress(dataset.sublist(6)),
+      : cvAddress = bigEndianData2cvAddress(dataset.sublist(6)),
         value = dataset[8] {
     assert(dataset.length == 0x0A);
   }
@@ -494,7 +514,7 @@ class LanXLocoInfo implements Command {
   });
 
   LanXLocoInfo.fromDataset(Uint8List dataset)
-      : locoAddress = data2locoAddress(dataset.sublist(5)),
+      : locoAddress = bigEndianData2locoAddress(dataset.sublist(5)),
         mode = dataset[7] & 0x07,
         busy = dataset[7] & (1 << 3) == (1 << 3),
         speedSteps = dataset[7] & 0x07,